pyproject.toml

[tool.poetry]
name = "ecom_scrapper"
version = "0.0.1"
description = ""
authors = ["Jose Alejandro Montana"]
readme = "README.md"
packages = [
        {include = "ecom_scrapper",from="src"},
        {include = "tests"}
]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.3"
numpy = ">=1.26.4,<2"
colorlog = "^6.9.0"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
black = "^23.12.1"
flake8 = "^7.0.0"
isort = "^5.13.2"
pydocstyle = "^6.3.0"
pytest = ">=8.2"
pylint = "^3.0.3"
coverage = "^7.4.0"
pytest-cov = "^5.0.0"
ipykernel = "^6.29.5"

[tool.black]
line-length = 120
include = '.*\.py$'
exclude = '''
(
  | notebooks
)
'''

[tool.pytest.ini_options]
minversion = "2.0"
addopts = "-rfEX -p pytester --strict-markers"
python_files = ["test_*.py", "*_test.py", "tests/*.py"]
testpaths = ["tests"]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"

[tool.isort]
profile = 'black'
skip = ["notebooks"]

[tool.pydocstyle]
convention = "google"
add-ignore = "D100,D101,D102,D103,D104,D105,D106,D107"

[tool.ruff]
line-length = 120
exclude = ["notebooks"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
